Use Login component to handle token collection
