+++ infrastructure/files/stylo.huma-num.fr.conf
@@ -44,7 +44,7 @@ server {
     proxy_redirect off;
   }
 
-  location ~ ^/(graphql|login|login/openid|logout|profile|authorization-code) {
+  location ~ ^/(graphql|login/openid|login/local|login/zotero|logout|authorization-code) {
     proxy_set_header X-Real-IP $remote_addr;
     proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
     proxy_set_header X-Forwarded-Proto $scheme;
